refactor:  modal components for app and user management - create base Modal component to wrap different modal content - move "container" components into components/modals.
