+++ src/components/Modal/index.ts
@@ -1,0 +1,6 @@
+export { default as CreateAppModal } from './CreateAppModal';
+export { default as CreateUserModal } from './CreateUserModal';
+export { default as Modal } from './Modal';
+export { default as PasswordGeneratedModal } from './PasswordGeneratedModal';
+export { default as UpdateAppModal } from './UpdateAppModal';
+export { default as UpdateUserModal } from './UpdateUserModal';
