--- src/containers/index.ts
@@ -1,2 +1,0 @@
-export { CreateUserModal } from '../containers/CreateUserModal';
-export { UpdateUserModal } from '../containers/UpdateUserModal';
